requirements.txt
# NotHere.one Crawler Dependencies

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Database
psycopg2-binary>=2.9.9

# Cache/Queue
redis>=5.0.0

# Optional but recommended
python-dotenv>=1.0.0  # For .env file support